BAEK/feel-log-app/backend/requirements.txt
flask 
flask-cors 
psycopg2-binary 
bcrypt
python-dotenv